refactor : global 패키지 추가
