+++ src/main/java/nextstep/domain/reservation/repository/ReservationMemoryRepository.java
@@ -1,7 +1,7 @@
 package nextstep.domain.reservation.repository;
 
 import nextstep.domain.reservation.domain.Reservation;
-import nextstep.exceptions.exception.ReservationNotFoundException;
+import nextstep.global.exceptions.exception.ReservationNotFoundException;
 import org.springframework.stereotype.Repository;
 
 import java.time.LocalDate;
