+++ src/main/java/nextstep/global/exceptions/exception/DuplicatedDateAndTimeException.java
@@ -1,4 +1,4 @@
-package nextstep.exceptions.exception;
+package nextstep.global.exceptions.exception;
 
 public class DuplicatedDateAndTimeException extends RuntimeException {
 }
