+++ src/main/java/nextstep/global/exceptions/exception/ReservationNotFoundException.java
@@ -1,4 +1,4 @@
-package nextstep.exceptions.exception;
+package nextstep.global.exceptions.exception;
 
 public class ReservationNotFoundException extends RuntimeException {
 }
